% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_fig_accessible.R
\name{exp_fig_accessible}
\alias{exp_fig_accessible}
\title{Export figure and alternative text}
\usage{
exp_fig_accessible(figure, alt_text, path, width = 5, height = 5, units = "cm")
}
\arguments{
\item{figure}{Plot object in the R environment to export.}

\item{alt_text}{Alternative text for the figure.}

\item{path}{Directory in which "exported" folder should be saved.}

\item{width}{Plot width, in units (see argument below). Default
is 5.}

\item{height}{Plot height, in units (see argument below). Default
is 5.}

\item{units}{Plot size units (options: "in", "cm", "mm", "px").
Default is "cm".}
}
\value{
A folder containing exported plots and associated
 alternative text as .png and .csv objects, respectively.
}
\description{
Function to export a specified plot object, and its alternative
text, from the R environment.
}
\examples{
\dontrun{
model_data <- utils::read.csv(here::here("data/Petrale_sole_std_res_2023.csv"))

fig1 <- satf::plot_spawning_biomass(model_data)

exp_fig_accessible(figure = fig1,
                   alt_text = "my alt text",
                   path = here::here(),
                   width = 5,
                   height = 5,
                   units = "cm")

}
}
