% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hake}
\alias{hake}
\title{Example `dat` object using Big Hake}
\format{
## `who`
A data frame with 26,4713 rows and 31 columns:
\describe{
  \item{label} {}
  \item{estimate} {}
  \item{year} {}
  \item{fleet} {}
  \item{sex} {}
  \item{area} {}
  \item{growth_pattern} {}
  \item{module_name} {}
  \item{uncertainty_label} {}
  \item{uncertainty} {}
  \item{time} {}
  \item{season} {}
  \item{subseason} {}
  \item{age} {}
  \item{len_bins} {}
  \item{initial} {}
  \item{likelihood} {}
  \item{gradient} {}
  \item{estimated} {}
  \item{bio_pattern} {}
  \item{birthseas} {}
  \item{settlement} {}
  \item{morph} {}
  \item{type} {}
  \item{factor} {}
  \item{platoon} {}
  \item{month} {}
  \item{sexes} {}
  \item{part} {}
  \item{bin} {}
  \item{kind} {}
}
}
\source{
<https://www.who.int/teams/global-tuberculosis-programme/data>
}
\usage{
hake
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
