% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_biomass.R
\name{plot_biomass}
\alias{plot_biomass}
\title{Plot Total Biomass}
\usage{
plot_biomass(
  dat,
  unit_label = "metric tons",
  scale_amount = 1,
  ref_line = c("target", "MSY", "msy", "unfished"),
  ref_point = NULL,
  end_year = NULL,
  relative = FALSE,
  make_rda = FALSE,
  rda_dir = getwd()
)
}
\arguments{
\item{dat}{A data frame returned from `asar::convert_output()`.}

\item{unit_label}{units for recruitment}

\item{scale_amount}{indicate the exact amount of scale (i.e. 1000)}

\item{ref_line}{A string specifying the type of reference you want to
compare biomass to. The default is `"target"`, which looks for
`"biomass_target"` in the `"label"` column of `dat`. The actual
searching in `dat` is case agnostic and will work with either upper- or
lower-case letters but you must use one of the options specified in the
default list to ensure that the label on the figure looks correct
regardless of how it is specified in `dat`.}

\item{ref_point}{A known value of the reference point along with the label
for the reference point as specified in the output file. Please use this
option if the ref_line cannot find your desired point. Indicate the
reference point in the form c("label" = value).}

\item{end_year}{last year of assessment}

\item{relative}{A logical value specifying if the resulting figures should
be relative spawning biomass. The default is `FALSE`. `ref_line` indicates
which reference point to use.}

\item{make_rda}{TRUE/FALSE; indicate whether to produce an .rda file containing
a list with the figure/table, caption, and alternative text (if figure). If TRUE,
the .rda will be exported to the folder indicated in the argument "rda_dir".
Default is FALSE.}

\item{rda_dir}{The location of the folder containing the generated .rda files
("rda_files") that will be created if the argument `make_rda` = TRUE.
Default is the working directory.}
}
\value{
Plot total biomass from a stock assessment model as found in a NOAA
stock assessment report. Units of total biomass can either be manually added
or will be extracted from the provided file if possible. In later releases, model will not
}
\description{
Plot Total Biomass
}
