% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_indices.R
\name{plot_indices}
\alias{plot_indices}
\title{Plot Index of Abundance}
\usage{
plot_indices(dat, unit_label = NULL, make_rda = TRUE, rda_dir = NULL)
}
\arguments{
\item{dat}{A data frame returned from `asar::convert_output()`.}

\item{unit_label}{units for index of abundance/CPUE}

\item{make_rda}{TRUE/FALSE; indicate whether to produce an .rda file containing
a list with the figure/table, caption, and alternative text (if figure). If TRUE,
the .rda will be exported to the folder indicated in the argument "rda_dir".
Default is FALSE.}

\item{rda_dir}{The location of the folder containing the generated .rda files
("rda_files") that will be created if the argument `make_rda` = TRUE.
Default is the working directory.}
}
\value{
Plot the estimated indices as indicated from a standard assessment
model output file.
}
\description{
Plot Index of Abundance
}
