% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recruitment.R
\name{plot_recruitment}
\alias{plot_recruitment}
\title{Plot Recruitment}
\usage{
plot_recruitment(
  dat,
  params = FALSE,
  params_only = FALSE,
  units = c(sb = "metric tons", recruitment = "metric tons"),
  recruitment_units = "metric tons",
  spawning_biomass_units = "metric tons",
  scaled = FALSE,
  scale_amount = NULL,
  show_warnings = FALSE,
  end_year = NULL,
  return = "recruitment",
  make_rda = FALSE,
<<<<<<< HEAD
<<<<<<< HEAD
  rda_dir = getwd()
=======
  rda_folder = getwd()
>>>>>>> 8fd15da (fix: minor fixes to get plot_recruitment and plot_spawning_biomass to work; used devtools::check, which updated .Rd files)
=======
  rda_dir = getwd()
>>>>>>> d7b90be (adding tests; also added libraries before read and write.csv fxns; updated documentation)
)
}
\arguments{
\item{dat}{A data frame returned from `asar::convert_output()`.}

\item{params}{Print/export the parameters of the stock recruitment function?}

\item{params_only}{Only export the stock recruitment function or both the parameters and the plot(s)?}

\item{units}{If units are not available in the output file, in metric tons,
or are different for SB and R, then report them here starting with SB units
and following with R units.}

\item{recruitment_units}{units for recruitment}

\item{spawning_biomass_units}{units of spawning biomass if different from biomass}

\item{scaled}{T/F; indicate whether the output values for biomass and recruitment are scaled}

\item{scale_amount}{indicate the exact amount of scale (i.e. 1000)}

\item{show_warnings}{Include warnings? Default FALSE}

\item{end_year}{last year of assessment}

\item{return}{Default returns recruitment over time. Options to display recruitment time series, stock recruitment curve, or recruitment fit}

<<<<<<< HEAD
<<<<<<< HEAD
\item{rda_dir}{The location of the folder containing the generated .rda files
=======
\item{rda_folder}{The location of the folder containing the generated .rda files
>>>>>>> 8fd15da (fix: minor fixes to get plot_recruitment and plot_spawning_biomass to work; used devtools::check, which updated .Rd files)
=======
\item{rda_dir}{The location of the folder containing the generated .rda files
>>>>>>> d7b90be (adding tests; also added libraries before read and write.csv fxns; updated documentation)
("rda_files") that will be created if the argument `export_rda` = TRUE.
Default is the working directory.}

\item{export_rda}{TRUE/FALSE; indcate whether to produce an .rda file containing
a list with the figure/table, caption, and alternative text (if figure). If TRUE,
<<<<<<< HEAD
<<<<<<< HEAD
the .rda will be exported to the folder indicated in the argument "rda_dir".
=======
the .rda will be exported to the folder indicated in the argument "rda_folder".
>>>>>>> 8fd15da (fix: minor fixes to get plot_recruitment and plot_spawning_biomass to work; used devtools::check, which updated .Rd files)
=======
the .rda will be exported to the folder indicated in the argument "rda_dir".
>>>>>>> d7b90be (adding tests; also added libraries before read and write.csv fxns; updated documentation)
Default is FALSE.}
}
\value{
A series of plots are exported including recruitment over time with R0
reference line, stock recruitment curve, and other related figures.
}
\description{
Plot Recruitment
}
