% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recruitment_deviations.R
\name{plot_recruitment_deviations}
\alias{plot_recruitment_deviations}
\title{Plot Recruitment Deviations}
\usage{
plot_recruitment_deviations(
  dat = NULL,
  end_year = NULL,
  n_projected_years = 10,
  make_rda = FALSE,
  rda_dir = getwd()
)
}
\arguments{
\item{dat}{A data frame returned from `asar::convert_output()`.}

\item{end_year}{last year of assessment}

\item{n_projected_years}{Number of years spawning biomass is projected for.
By default this number is set to 10}

\item{make_rda}{TRUE/FALSE; indicate whether to produce an .rda file containing
a list with the figure/table, caption, and alternative text (if figure). If TRUE,
the .rda will be exported to the folder indicated in the argument "rda_dir".
Default is FALSE.}

\item{rda_dir}{The location of the folder containing the generated .rda files
("rda_files") that will be created if the argument `make_rda` = TRUE.
Default is the working directory.}
}
\value{
Plot recruitment deviations relative to one over time from an
assessment model output file translated to a standardized output. There are
options to return a {ggplot2} object or export an .rda object containing
associated caption and alternative text for the figure.
}
\description{
Plot Recruitment Deviations
}
