% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spawning_biomass.R
\name{plot_spawning_biomass}
\alias{plot_spawning_biomass}
\title{Plot Spawning Biomass}
\usage{
plot_spawning_biomass(
  dat,
  model = c("SS3", "BAM", "ASAP", "AMAK"),
  refpts = FALSE,
  warnings = FALSE,
  units = NULL
)
}
\arguments{
\item{dat}{Stock assessment output file containing estimates parameters and
other associated metrics.}

\item{model}{Acryonym of the stock assessment model that produced the output file.
Currently, only stock synthesis (SS3) and Beaufort Assessment Model (BAM)
are options for plotting spawning biomass.}

\item{refpts}{Output a dataframe of reference points for spawning biomass
rather than plot. Default is FALSE.}

\item{warnings}{Option to suppress warnings}

\item{units}{If units are not available in the output file or are not the
default of metric tons, then state the units of spawning biomass applicable
to the stock. For plotting purposes, spawning biomass is divided by 1000.}
}
\value{
Plot spawning biomass from a stock assessment model as found in a NOAA
stock assessment report. Units of spawning biomass can either be manually added
or will be extracted from the provided file if possible.
}
\description{
Plot Spawning Biomass
}
\examples{
plot_spawning_biomass(dat = "spp.rdat",model = "BAM")
}
