% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_biomass.R
\name{plot_total_biomass}
\alias{plot_total_biomass}
\title{Plot Total Biomass}
\usage{
plot_total_biomass(
  dat,
  show_warnings = FALSE,
  units = NULL,
  scaled = FALSE,
  scale_amount = 1000,
  ref_line = c("target", "MSY", "msy", "unfished"),
  end_year = NULL,
  relative = FALSE,
  make_rda = FALSE,
  rda_folder = getwd()
)
}
\arguments{
\item{dat}{A data frame returned from `asar::convert_output()`.}

\item{show_warnings}{Option to suppress warnings}

\item{units}{units for biomass}

\item{scaled}{TRUE/FALSE; indicate whether the output values for biomass and recruitment are scaled}

\item{scale_amount}{indicate the exact amount of scale (i.e. 1000)}

\item{ref_line}{choose with reference point to plot a reference line and use
in relative totb calculations}

\item{end_year}{input the end year of the stock assessment data (not including
projections). This parameter will be deprecated once the output converter is fully developed.}

\item{relative}{Plot relative total biomass. Ref line indicates which reference point to use}

\item{rda_folder}{The location of the folder containing the generated .rda files
("rda_files") that will be created if the argument `export_rda` = TRUE.
Default is the working directory.}
}
\value{
Plot total biomass from a stock assessment model as found in a NOAA
stock assessment report. Units of total biomass can either be manually added
or will be extracted from the provided file if possible. In later releases, model will not
}
\description{
Plot Total Biomass
}
