% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_bnc.R
\name{table_bnc}
\alias{table_bnc}
\title{Biomass, abundance, and catch time series table}
\usage{
table_bnc(
  dat,
  end_year = NULL,
  biomass_unit_label = "mt",
  catch_unit_label = "mt",
  make_rda = FALSE,
  rda_dir = getwd()
)
}
\arguments{
\item{dat}{A data frame returned from `asar::convert_output()`.}

\item{end_year}{last year of assessment}

\item{biomass_unit_label}{abbreviated units for biomass}

\item{catch_unit_label}{abbreviated units for catch}

\item{make_rda}{TRUE/FALSE; indicate whether to produce an .rda file containing
a list with the figure/table, caption, and alternative text (if figure). If TRUE,
the .rda will be exported to the folder indicated in the argument "rda_dir".
Default is FALSE.}

\item{rda_dir}{The location of the folder containing the generated .rda files
("rda_files") that will be created if the argument `make_rda` = TRUE.
Default is the working directory.}
}
\value{
Create a table of biomass, abundance, and catch through all years of
the assessment model output translated to a standard structure.There are
options to return a {flextable} object or export an rda object containing
associated caption for the table.
}
\description{
Biomass, abundance, and catch time series table
}
