% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_indices.R
\name{table_indices}
\alias{table_indices}
\title{Create Indices of Abundance Table}
\usage{
table_indices(dat)
}
\arguments{
\item{dat}{A data frame returned from `asar::convert_output()`.}
}
\value{
Create table of observed annual indices of abundance plus error
stratified by fleet.
}
\description{
Create Indices of Abundance Table
}
