% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_all_figs_tables.R
\name{exp_all_figs_tables}
\alias{exp_all_figs_tables}
\title{Export all figures and tables}
\usage{
exp_all_figs_tables(
  dat,
  unit_label = "metric tons",
  scale_amount = 1,
  end_year = NULL,
  n_projected_years = 10,
  relative = FALSE,
  rda_dir = getwd(),
  ref_line = c("target", "MSY", "msy", "unfished"),
  spawning_biomass_label = "metric tons",
  recruitment_label = "metric tons",
  ref_line_sb = c("target", "MSY", "msy", "unfished")
)
}
\arguments{
\item{dat}{A data frame returned from `asar::convert_output()`.}

\item{unit_label}{units for recruitment}

\item{scale_amount}{indicate the exact amount of scale (i.e. 1000)}

\item{end_year}{last year of assessment}

\item{n_projected_years}{Number of years spawning biomass is projected for.
By default this number is set to 10}

\item{relative}{A logical value specifying if the resulting figures should
be relative spawning biomass. The default is `FALSE`. `ref_line` indicates
which reference point to use.}

\item{make_rda}{TRUE/FALSE; indicate whether to produce an .rda file containing
a list with the figure/table, caption, and alternative text (if figure). If TRUE,
the .rda will be exported to the folder indicated in the argument "rda_dir".
Default is FALSE.}

\item{rda_dir}{The location of the folder containing the generated .rda files
("rda_files") that will be created if the argument `make_rda` = TRUE.
Default is the working directory.}

\item{ref_line}{@inheritParams plot_biomass}

\item{spawning_biomass_label}{@inheritParams plot_spawn_recruitment}

\item{recruitment_label}{@inheritParams plot_spawn_recruitment}

\item{ref_line_sb}{Identical definition as `ref_line`, but this argument is
applied to plot_spawning_biomass.}
}

\description{
Export all figures and tables to Rda files within one function.
}
\examples{
\dontrun{
exp_all_figs_tables(dat, ref_line = "unfished",
ref_line_sb = "target", make_rda = TRUE)
}
}
