% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_landings.R
\name{plot_landings}
\alias{plot_landings}
\title{Plot observed landings by fleet}
\usage{
plot_landings(
  dat,
  unit_label = "metric tons",
  make_rda = FALSE,
  rda_dir = getwd()
)
}
\arguments{
\item{dat}{A data frame returned from `asar::convert_output()`.}

\item{unit_label}{indicate the name of the units of landings as to label the axis}

\item{make_rda}{TRUE/FALSE; indicate whether to produce an .rda file containing
a list with the figure/table, caption, and alternative text (if figure). If TRUE,
the .rda will be exported to the folder indicated in the argument "rda_dir".
Default is FALSE.}

\item{rda_dir}{The location of the folder containing the generated .rda files
("rda_files") that will be created if the argument `make_rda` = TRUE.
Default is the working directory.}
}
\value{
Create a plot ready for a stock assessment report of cumulative landings
over time by fleet.Includes options to plot by fleet, total observed landings
with and without predicted landings. Indicate if fleet should be faceted or on one plot (default). Warning: i
}
\description{
Plot observed landings by fleet
}
