% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recruitment.R
\name{plot_recruitment}
\alias{plot_recruitment}
\title{Plot Recruitment}
\usage{
plot_recruitment(
  dat,
  unit_label = "metric tons",
  scale_amount = 1,
  end_year = NULL,
  n_projected_years = 10,
  relative = FALSE
)
}
\arguments{
\item{dat}{A data frame returned from `asar::convert_output()`.}

\item{unit_label}{units for recruitment}

\item{scale_amount}{indicate the exact amount of scale (i.e. 1000)}

\item{end_year}{last year of assessment}

\item{n_projected_years}{Number of years spawning biomass is projected for.
By default this number is set to 10}

\item{relative}{A logical value specifying if the resulting figures should
be relative spawning biomass. The default is `FALSE`. `ref_line` indicates
which reference point to use.}
}
\value{
Plot recruitment over time from an assessment model output file
translated to a standardized output. There are options to return a {ggplot2}
object or export an rda object containing associated caption and alternative
text for the figure.
}
\description{
Plot Recruitment
}
