% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recruitment.R
\name{plot_recruitment}
\alias{plot_recruitment}
\title{Plot Recruitment}
\usage{
plot_recruitment(
  dat,
  model,
  params = FALSE,
  params_only = FALSE,
  units = NULL,
  include_fxn = FALSE,
  warnings = FALSE
)
}
\arguments{
\item{dat}{Stock assessment output file containing estimates parameters and
other associated metrics.}

\item{model}{Acryonym of the stock assessment model that produced the output file.
Currently, only stock synthesis (SS3) and Beaufort Assessment Model (BAM)
are options for plotting spawning biomass.}

\item{params}{Print/export the parameters of the stock recruitment function?}

\item{params_only}{Only export the stock recruitment function or both the parameters and the plot(s)?}

\item{units}{If units are not available in the output file, in metric tons,
or are different for SB and R, then report them here starting with SB units
and following with R units.}

\item{include_fxn}{Include the stock recruitment function on the SR curve plot?}

\item{warnings}{Include warnings? Default FALSE}
}
\value{
A series of plots are exported including recruitment over time with R0
reference line, stock recruitment curve, and other related figures.
}
\description{
Plot Recruitment
}
