% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recruitment.R
\name{plot_recruitment}
\alias{plot_recruitment}
\title{Plot Recruitment}
\usage{
plot_recruitment(
  dat,
  params = FALSE,
  params_only = FALSE,
  units = c(sb = "metric tons", recruitment = "metric tons"),
  recruitment_units = "metric tons",
  spawning_biomass_units = "metric tons",
  scaled = FALSE,
  scale_amount = NULL,
  show_warnings = FALSE,
  end_year = NULL,
  return = "recruitment"
)
}
\arguments{
\item{dat}{A data frame returned from `asar::convert_output()`.}

\item{params}{Print/export the parameters of the stock recruitment function?}

\item{params_only}{Only export the stock recruitment function or both the parameters and the plot(s)?}

\item{units}{If units are not available in the output file, in metric tons,
or are different for SB and R, then report them here starting with SB units
and following with R units.}

\item{recruitment_units}{units for recruitment}

\item{spawning_biomass_units}{units of spawning biomass if different from biomass}

\item{scaled}{T/F; indicate whether the output values for biomass and recruitment are scaled}

\item{scale_amount}{indicate the exact amount of scale (i.e. 1000)}

\item{show_warnings}{Include warnings? Default FALSE}

\item{end_year}{last year of assessment}

\item{return}{Default returns recruitment over time. Options to display recruitment time series, stock recruitment curve, or recruitment fit}
}
\value{
A series of plots are exported including recruitment over time with R0
reference line, stock recruitment curve, and other related figures.
}
\description{
Plot Recruitment
}
