% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recruitment_deviations.R
\name{plot_recruitment_deviations}
\alias{plot_recruitment_deviations}
\title{Plot Recruitment Deviations}
\usage{
plot_recruitment_deviations(
  dat = NULL,
  end_year = NULL,
  n_projected_years = 10,
  make_rda = FALSE,
  rda_dir = getwd()
)
}
\value{
Plot recruitment deviations relative to one over time from an
assessment model output file translated to a standardized output. There are
options to return a {ggplot2} object or export an .rda object containing
associated caption and alternative text for the figure.
}
\description{
@inheritParams plot_recruitment
}
