% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spawn_recruitment.R
\name{plot_spawn_recruitment}
\alias{plot_spawn_recruitment}
\title{Plot Spawn-Recruit Curve}
\usage{
plot_spawn_recruitment(
  dat = NULL,
  spawning_biomass_label = "metric tons",
  recruitment_label = "metric tons",
  end_year = NULL,
  make_rda = FALSE,
  rda_dir = getwd()
)
}
\value{
Plot spawning recruitment relationship from a standardized output file originating from asar::convert_output()
}
\description{
@inheritParams plot_recruitment
 @param spawning_biomass_label Units for spawning biomass
 @param recruitment_label units for recruitment
}
