% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spawning_biomass.R
\name{plot_spawning_biomass}
\alias{plot_spawning_biomass}
\title{Plot Spawning Biomass}
\usage{
plot_spawning_biomass(
  dat,
  model = "standard",
  show_warnings = FALSE,
  units = NULL,
  spawning_biomass_units = NULL,
  scaled = FALSE,
  scale_amount = 1000,
  ref_line = c("target", "MSY", "msy", "unfished"),
  end_year = NULL,
  relative = FALSE
)
}
\arguments{
\item{dat}{Stock assessment output file containing estimates parameters and
other associated metrics.}

\item{model}{Acryonym of the stock assessment model that produced the output file.
Currently, only stock synthesis (SS3) and Beaufort Assessment Model (BAM)
are options for plotting spawning biomass.}

\item{show_warnings}{Option to suppress warnings}

\item{units}{If units are not available in the output file, in metric tons,
or are different for SB and R, then report them here starting with SB units
and following with R units.}

\item{spawning_biomass_units}{units of spawning biomass if different from biomass}

\item{scaled}{T/F; indicate whether the output values for biomass and recruitment are scaled}

\item{scale_amount}{indicate the exact amount of scale (i.e. 1000)}

\item{ref_line}{choose with reference point to plot a reference line and use
in relative sb calculations}

\item{end_year}{input the end year of the stock assessment data (not including
projections). This parameter will be deprecated once the output converter is fully developed.}

\item{relative}{Plot relative spawning biomass. Ref line indicates which reference point to use}

\item{biomass_units}{units for biomass}
}
\value{
Plot spawning biomass from a stock assessment model as found in a NOAA
stock assessment report. Units of spawning biomass can either be manually added
or will be extracted from the provided file if possible. In later releases, model will not
}
\description{
Plot Spawning Biomass
}
