% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spawning_biomass.R
\name{plot_spawning_biomass}
\alias{plot_spawning_biomass}
\title{Plot spawning biomass (SB)}
\usage{
plot_spawning_biomass(
  dat,
  unit_label = "metric ton",
  scale_amount = 1,
  ref_line = c("target", "unfished"),
  end_year = NULL,
  relative = FALSE,
  make_rda = FALSE,
  rda_dir = getwd()
)
}
\arguments{
\item{dat}{A data frame returned from `asar::convert_output()`.}

\item{scale_amount}{indicate the exact amount of scale (i.e. 1000)}

\item{ref_line}{A string specifying the type of reference you want to
compare spawning biomass to. The default is `"target"`, which looks for
`"spawning_biomass_target"` in the `"label"` column of `dat`. The actual
searching in `dat` is case agnostic and will work with either upper- or
lower-case letters but you must use one of the options specified in the
default list to ensure that the label on the figure looks correct
regardless of how it is specified in `dat`.}

\item{end_year}{last year of assessment}

\item{relative}{A logical value specifying if the resulting figures should
be relative spawning biomass. The default is `FALSE`. `ref_line` indicates
which reference point to use.}

\item{rda_dir}{The location of the folder containing the generated .rda files
("rda_files") that will be created if the argument `export_rda` = TRUE.
Default is the working directory.}
}
\value{
Plot spawning biomass from the results of an assessment model translated to
the standard output. The {ggplot2} object is returned for further
modifications if needed.
}
\description{
Plot spawning biomass with a reference line as a dashed line. The figure can
also be made relative to this reference line rather than in absolute units.
}
