% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_indices.R
\name{table_indices}
\alias{table_indices}
\title{Create Indices of Abundance Table}
\usage{
table_indices(dat, model = "standard")
}
\arguments{
\item{dat}{.csv file and path where the data is located of standard stock assessment output}

\item{model}{the model in which the output came from (standard, SS3, BAM, ect)}
}
\value{
Create table of observed annual indices of abundance plus error
stratified by fleet.
}
\description{
Create Indices of Abundance Table
}
