% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_indices.R
\name{table_indices}
\alias{table_indices}
\title{Create Indices of Abundance Table}
\usage{
<<<<<<< HEAD
table_indices(dat, make_rda = FALSE, rda_dir = getwd())
=======
table_indices(dat, make_rda = FALSE, rda_folder = getwd())
>>>>>>> 8fd15da (fix: minor fixes to get plot_recruitment and plot_spawning_biomass to work; used devtools::check, which updated .Rd files)
}
\arguments{
\item{dat}{A data frame returned from `asar::convert_output()`.}

<<<<<<< HEAD
\item{rda_dir}{The location of the folder containing the generated .rda files
=======
\item{rda_folder}{The location of the folder containing the generated .rda files
>>>>>>> 8fd15da (fix: minor fixes to get plot_recruitment and plot_spawning_biomass to work; used devtools::check, which updated .Rd files)
("rda_files") that will be created if the argument `export_rda` = TRUE.
Default is the working directory.}
}
\value{
Create table of observed annual indices of abundance plus error
stratified by fleet.
}
\description{
Create Indices of Abundance Table
}
