% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_indices.R
\name{table_indices}
\alias{table_indices}
\title{Create Indices of Abundance Table}
\usage{
table_indices(dat, model = "standard")
}
\arguments{
\item{dat}{Stock assessment output file containing estimates parameters and
other associated metrics.}

\item{model}{Acryonym of the stock assessment model that produced the output file.
Currently, only stock synthesis (SS3) and Beaufort Assessment Model (BAM)
are options for plotting spawning biomass.}
}
\value{
Create table of observed annual indices of abundance plus error
stratified by fleet.
}
\description{
Create Indices of Abundance Table
}
