% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_captions.R
\name{write_captions}
\alias{write_captions}
\title{Write captions and alternative text}
\usage{
write_captions(dat, dir = NULL, year = NULL)
}
\arguments{
\item{dat}{A data frame returned from `asar::convert_output()`.}

\item{dir}{Directory where the output captions and alt text file should be saved}

\item{year}{the last year of the data or the current year this function is being performed}
}
\value{
Exports .csv with captions and alt text for figures and tables
that contain key quantities (e.g., an assessment's start year) that
are automatically extracted from the converted model results file.
}
\description{
Function to create captions and alternative text that contain
key quantities from the model results file.
}
